Create Query 1 Orders table creation
This query is the basic query used to join the relevant Key fields on the orders table and answer basic questions such as How many orders? Average order value etc. The purpose of this query is mainly just to begin drawing the information together on its main table and select the relevant columns.  This is done by first selecting the table name(specified in the pre planning/table creation phase) then joining on the relevant fields. The lesser significant measures are done in the reporting tool.