Insertion code examples.
This is a sample of the insertion data. The data of course will not be shown in full here but this is an example of how the data was inserted. Issues arose with MYSQL syntax needing to be transposed to postgresql syntax and some data/tables had to be re arranged to fit the source data. Name timestamp fields in postgresql and the datetime values did not fit. Postgres does timestamp with or without time zone. Ultimately decided to change the column data type to just 'time'.